Добавляет фильтры в каждый контроллер отдельно
