+++ frontend/controllers/UsersController.php
@@ -4,12 +4,12 @@ declare(strict_types = 1);
 
 namespace frontend\controllers;
 
-use frontend\controllers\SecuredController;
+use yii\web\Controller;
 use frontend\models\user\UserService;
 use frontend\models\user\UserSearchForm;
 
 
-class UsersController extends SecuredController
+class UsersController extends Controller
 {
     private UserService $service;
 
@@ -17,6 +17,22 @@ class UsersController extends SecuredController
     {
         parent::init();
         $this->service = new UserService($this->request);
+    }
+
+    public function behaviors()
+    {
+        return [
+            'access' => [
+                'class' => AccessControl::class,
+                'rules' => [
+                    [
+                        'actions' => ['index', 'view'],
+                        'allow' => true,
+                        'roles' => ['@']
+                    ]
+                ]
+            ]
+        ];
     }
 
     public function actionIndex()
